+++ src/simulation/api/agent.controller.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
-import agentService from '../service/agent.service';
+import agentService from '@simulation/service/agent.service';
 import { AgentDocument } from '@simulation/db/models/agent.model';
-import { INTERNAL_SERVER_ERROR } from '../utils/errors';
+import { INTERNAL_SERVER_ERROR } from '@simulation/utils/errors';
 
 /**
  * Creates a new agent.
