+++ src/simulation/api/auth.controller.ts
@@ -1,7 +1,7 @@
 /* eslint-disable require-jsdoc */
 // Controller that implements authentication
 import { Request, Response } from 'express';
-import { INTERNAL_SERVER_ERROR } from '../utils/errors';
+import { INTERNAL_SERVER_ERROR } from '@simulation/utils/errors';
 
 async function login(req: Request, res: Response): Promise<void> {
   const { accessCode } = req.body;
