+++ src/simulation/api/chat.controller.ts
@@ -1,10 +1,10 @@
 // Controller that implements chat related endpoints
 import { Request, Response } from 'express';
-import { SimulationDocument } from '../db/models/simulation.model';
+import { SimulationDocument } from '@simulation/db/models/simulation.model';
 
-import chatService from '../service/chat.service';
-import { INTERNAL_SERVER_ERROR } from '../utils/errors';
-import ChatMessage from '../model/response/chat.response';
+import chatService from '@simulation/service/chat.service';
+import { INTERNAL_SERVER_ERROR } from '@simulation/utils/errors';
+import ChatMessage from '@simulation/model/response/chat.response';
 import { StartChatRequest } from '@simulation/model/request/chat.request';
 
 /**
