+++ src/simulation/api/optimization.controller.ts
@@ -2,12 +2,12 @@
 import { Request, Response } from 'express';
 import { validationResult } from 'express-validator';
 
-import { RunSimulationRequest } from '../model/request/simulation.request';
-import optimizationService from '../service/optimization.service';
-import { logger } from '../service/logging.service';
+import { RunSimulationRequest } from '@simulation/model/request/simulation.request';
+import optimizationService from '@simulation/service/optimization.service';
+import { logger } from '@simulation/service/logging.service';
 
-import { INTERNAL_SERVER_ERROR } from '../utils/errors';
-import { OptimizationDocument } from '../db/models/optimization.model';
+import { INTERNAL_SERVER_ERROR } from '@simulation/utils/errors';
+import { OptimizationDocument } from '@simulation/db/models/optimization.model';
 import { SimulationDocument } from '@simulation/db/models/simulation.model';
 
 /**
