+++ src/simulation/api/simulation.controller.ts
@@ -2,15 +2,15 @@
 import { Request, Response } from 'express';
 import { validationResult } from 'express-validator';
 
-import { RunABTestingRequest } from '../model/request/run-ab-testing.request';
-import { RunSimulationRequest } from '../model/request/simulation.request';
+import { RunABTestingRequest } from '@simulation/model/request/run-ab-testing.request';
+import { RunSimulationRequest } from '@simulation/model/request/simulation.request';
 
-import { SimulationDocument } from '../db/models/simulation.model';
-import simulationService from '../service/simulation.service';
-import { logger } from '../service/logging.service';
-import { ConversationDocument } from '../db/models/conversation.model';
+import { SimulationDocument } from '@simulation/db/models/simulation.model';
+import simulationService from '@simulation/service/simulation.service';
+import { logger } from '@simulation/service/logging.service';
+import { ConversationDocument } from '@simulation/db/models/conversation.model';
 
-import { INTERNAL_SERVER_ERROR } from '../utils/errors';
+import { INTERNAL_SERVER_ERROR } from '@simulation/utils/errors';
 
 /**
  * Runs the simulation.
