+++ src/simulation/db/repositories/agent.repository.ts
@@ -1,6 +1,6 @@
 import { Model } from 'mongoose';
-import { BaseRepository } from './base.repository';
-import { AgentDocument } from '../models/agent.model';
+import { BaseRepository } from '@simulation/db/repositories/base.repository';
+import { AgentDocument } from '@simulation/db/models/agent.model';
 
 class AgentRepository extends BaseRepository<AgentDocument> {
   constructor(model: Model<AgentDocument>) {
