+++ src/simulation/model/request/run-ab-testing.request.ts
@@ -1,4 +1,4 @@
-import { SimulationType, SimulationScenario } from '../../db/enum/enums';
+import { SimulationType, SimulationScenario } from '@simulation/db/enum/enums';
 import { AgentDocument } from '@simulation/db/models/agent.model';
 
 interface RunABTestingRequest {
