+++ src/simulation/model/request/simulation.request.ts
@@ -1,5 +1,5 @@
-import { SimulationType, SimulationScenario } from '../../db/enum/enums';
-import { AgentDocument } from '../../db/models/agent.model';
+import { SimulationType, SimulationScenario } from '@simulation/db/enum/enums';
+import { AgentDocument } from '@simulation/db/models/agent.model';
 
 interface RunSimulationRequest {
   scenario: SimulationScenario;
