+++ src/simulation/model/response/dashboard.response.ts
@@ -1,4 +1,4 @@
-import { SimulationDocument } from '../../db/models/simulation.model';
+import { SimulationDocument } from '@simulation/db/models/simulation.model';
 
 interface DashboardData {
   interactions: number;
