+++ src/simulation/service/conversation.service.ts
@@ -1,8 +1,7 @@
 /* eslint-disable require-jsdoc */
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { SimulationDocument } from '../db/models/simulation.model';
-import { CustomAgent } from '../agents/custom.agent';
-import { getSimConfig } from '../agents/user.agent';
+import { CustomAgent } from '@simulation/agents/custom.agent';
+import { getSimConfig } from '@simulation/agents/user.agent';
 import { BaseChatModel, BaseChatModelParams } from 'langchain/chat_models/base';
 import { ChatOpenAI, OpenAIChatInput } from 'langchain/chat_models/openai';
 import { AzureOpenAIInput } from 'langchain/chat_models/openai';
@@ -10,15 +9,15 @@ import {
   flightBookingAgentConfig,
   fakeUserAgentResponses,
   fakeServiceAgentResponses,
-} from '../agents/service/service.agent.flight.booker';
+} from '@simulation/agents/service/service.agent.flight.booker';
 import { Types } from 'mongoose';
-import { MsgHistoryItem } from '../agents/custom.agent';
-import { AgentDocument, AgentModel } from '../db/models/agent.model';
+import { MsgHistoryItem } from '@simulation/agents/custom.agent';
+import { AgentDocument } from '@simulation/db/models/agent.model';
 import { FakeListChatModel } from 'langchain/chat_models/fake';
-import { LLMModel } from '../db/enum/enums';
-import repositoryFactory from '../db/repositories/factory';
-import { MessageDocument } from '../db/models/message.model';
-import { MsgSender, MsgTypes, ConversationStatus } from '../db/enum/enums';
+import { LLMModel } from '@simulation/db/enum/enums';
+import repositoryFactory from '@simulation/db/repositories/factory';
+import { MessageDocument } from '@simulation/db/models/message.model';
+import { MsgSender, MsgTypes, ConversationStatus } from '@simulation/db/enum/enums';
 import * as fs from 'fs';
 import * as path from 'path';
