+++ src/simulation/validator/simulation.validator.ts
@@ -1,5 +1,5 @@
 import { body, param, ValidationChain } from 'express-validator';
-import { SimulationType, SimulationScenario } from '../db/enum/enums';
+import { SimulationType, SimulationScenario } from '@simulation/db/enum/enums';
 
 class SimulationValidator {
   /**
