+++ tsconfig.json
@@ -34,10 +34,20 @@
       "@simulation/config/*": ["simulation/config/*"],
       "@simulation/db/*": ["simulation/db/*"],
       "@simulation/docs/*": ["simulation/docs/*"],
+      "@simulation/mockedAPI/*": ["simulation/mockedAPI/*"],
       "@simulation/model/*": ["simulation/model/*"],
       "@simulation/router/*": ["simulation/router/*"],
       "@simulation/service/*": ["simulation/service/*"],
-      "@simulation/utils/*": ["simulation/utils/*"]
+      "@simulation/utils/*": ["simulation/utils/*"],
+      "@simulation/validator/*": ["simulation/validator/*"],
+      "@evaluation/api/*": ["evaluation/api/*"],
+      "@evaluation/db/*": ["evaluation/db/*"],
+      "@evaluation/docs/*": ["evaluation/docs/*"],
+      "@evaluation/model/*": ["evaluation/model/*"],
+      "@evaluation/router/*": ["evaluation/router/*"],
+      "@evaluation/service/*": ["evaluation/service/*"],
+      "@evaluation/utils/*": ["evaluation/utils/*"],
+      "@evaluation/validator/*": ["evaluation/validator/*"]
     },
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
